+++ mirrord/kube/src/api/runtime.rs
@@ -16,7 +16,10 @@ use kube::{api::ListParams, Api, Client};
 use mirrord_config::target::{DeploymentTarget, PodTarget, RolloutTarget, Target};
 
 use crate::{
-    api::{container::choose_container, get_k8s_resource_api, kubernetes::rollout::Rollout},
+    api::{
+        container::choose_container,
+        kubernetes::{get_k8s_resource_api, rollout::Rollout},
+    },
     error::{KubeApiError, Result},
 };
